% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CI.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate the confidence intervals}
\usage{
calculate_CI(x, conf = 0.95)
}
\arguments{
\item{x}{a vector of numbers of length N}
}
\value{
the sample CI of x
}
\description{
Calculate the confidence intervals
}
\details{
This function calculate the confidence intervals
}
\examples{
a <- c(1,2,3,4,5)
calculate_CI(a)

}
