% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chisq_f.R
\name{Chisq_f}
\alias{Chisq_f}
\title{Chi-square Goodness-of-Fit Test for a sample of the Pf Distribution}
\usage{
Chisq_f(sample)
}
\arguments{
\item{sample}{Numeric vector of observations.}
}
\value{
A named numeric vector with:
\itemize{
  \item statistic : Chi-square statistic
  \item df        : Degrees of freedom
  \item p.val     : p-value
}
}
\description{
Performs a chi-square goodness-of-fit test comparing observed frequencies
from a sample with expected probabilities derived from the distribution:

\deqn{
f(x) =
\begin{cases}
\tfrac{3}{4} \, x \, (2 - x)^2, & 0 \le x \le 2 \\
0, & \text{otherwise}
\end{cases}
}

The interval [0, 2] is divided into bins of width 0.25.
}
\details{
This function tests the hypotheses \eqn{H_0: \pi = \hat{\pi}} vs. \eqn{H_1: \pi \neq \hat{\pi}}.
Expected probabilities are computed as differences of the CDF \eqn{F(x)} over each interval.
Observed frequencies in the sample are obtained using the \code{fdth} package.
}
\examples{
sample <- randomfun2(1000)
Chisq_f(sample)

}
