% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_plot.R
\name{coef_plot}
\alias{coef_plot}
\title{coef_plot}
\usage{
coef_plot(betahat, varn, plotname)
}
\arguments{
\item{betahat}{A vector that stores estimated coefficients.}

\item{varn}{A vector that stores names for the coefficients.}

\item{plotname}{A string that to be used as the plot name.}
}
\value{
A plot of coefficients.
}
\description{
Plotting the coefficients
}
\details{
This function plots the coefficients
\strong{Notice:} This function mainly serves as a functional part for other functions. So we seldomly
apply this function by itself.
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
