% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_harmonics.R
\name{ff_harmonics}
\alias{ff_harmonics}
\title{ff_harmonics}
\usage{
ff_harmonics(x = NULL, n = NULL, up = 10L, plot = TRUE, add = F,
  main = NULL, ...)
}
\arguments{
\item{x}{Original signal (1D, 2D, or 3D array).}

\item{n}{Number of first harmonics to report (integer).}

\item{up}{Upsamping rate (default=10).}

\item{plot}{Boolean indicating whether to print the harmonics plot(default==TRUE).}

\item{add}{whether to overplot the harmonics on an existing graph (default=FALSE),}

\item{main}{Title for the plot.}
}
\value{
A plot and a dataframe with the sampled harmonics and their corresponding FT magnitudes/amplitudes.
}
\description{
Compute the spectral decomposition of an array (harmonics)
}
\details{
This function computes the FT of the singlal and plots the first few harmonics
}
\examples{
ff_harmonics(x = y, n = 12L, up = 100L, col = 2L, lwd=3, cex=2)

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
