% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftinv.R
\name{fftinv}
\alias{fftinv}
\title{fftinv}
\usage{
fftinv(x)
}
\arguments{
\item{x}{FT of a dataset.}
}
\value{
The IFT of the input array.
}
\description{
Implicitly Invert the FT (IFT)
}
\details{
This function does the IFT and scales appropriately the  result to ensure IFT(FT()) = I()
}
\examples{
fftinv(fft(circle_arr))

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
