% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_ARIMA.R
\name{Fit_ARIMA}
\alias{Fit_ARIMA}
\title{Fit_ARIMA}
\usage{
Fit_ARIMA(country = "Belgium", start = 2000, end = 2017,
  frequency = 20,
  feature = "Unemployment , Females, From 15-64 years, Total")
}
\arguments{
\item{country}{Give a country name that is shown on the original dataset}

\item{start}{Select the start year to create the ARIMA model}

\item{end}{Select the end year to create the ARIMA model}

\item{frequency}{The number of observations per unit of time. The same in function "ts"}

\item{feature}{Choose one feature to create the ARIMA model on}
}
\description{
Fit the ARIMA model
}
\details{
This function cleans the data and fits the ARIMA model. It is based on several previous functions
and must work under the European Union Economy dataset. It is built primarily aiming to create ARIMA model for 
the EU Economy dataset.
}
\examples{
# Example are based on country 
BulgariaARIMA = 
Fit_ARIMA(country ='Bulgaria', 
         start=2000, end=2017, frequency=20,
         feature="Gross domestic product at market prices")
summary(BulgariaARIMA)
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
