% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin.R
\name{fn_sin}
\alias{fn_sin}
\title{Compute the sine}
\usage{
fn_sin(x, k)
}
\arguments{
\item{x}{numeric, the number to be transformed}

\item{k}{integer, the number of terms to be used in the series expansion beyond the constant 1. The value of k is always >= 1.}
}
\value{
the sine value of input number
}
\description{
Compute the sine of a number using the truncated series expansion
}
\details{
This function compute the sine of a number
}
\examples{
fn_sin(1,5)

}
