% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifftshift.R
\name{ifftshift}
\alias{ifftshift}
\title{ifftshift}
\usage{
ifftshift(img_ff, dim = -1)
}
\arguments{
\item{img_ff}{An Inverse Fourier transform of a 1D signal, 2D image, or 3D volume.}

\item{dim}{Number of dimensions (-1, 1, 2, 3).}
}
\value{
A properly shifted IFT of the input array.
}
\description{
IFFT SHIFT
}
\details{
This function is useful for moving back the zero-frequency component in the middle of the spectrum
back to (0,0,0).  It rearranges in reverse (relative to fftshift()) the indices appropriately,
so that the image can be correctly reconstructed by the IFT in spacetime
}
\examples{
ifftshift(fft(Real+1i*Imaginary, inverse = T))

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
