% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSpaceTransform.R
\name{kSpaceTransform}
\alias{kSpaceTransform}
\title{K-space Transformation}
\usage{
kSpaceTransform(data, inverse = FALSE, reconPhases = NULL)
}
\arguments{
\item{data}{dataset that needs K-space transformation}
}
\description{
This function is useful for transforming data to k-space.
}
\details{
A generic function to Transform Data ={all predictors (X) and outcome (Y)} to k-space (Fourier domain)
For ForwardFT, set parameters as (rawData, FALSE, NULL)
For InverseFT, there are two parameters setting: (magnitudes, TRUE, reconPhasesToUse) or (FT_data, TRUE, NULL)
}
\examples{
kSpaceTransform(mag_ft_xC_fMRI_train, TRUE, phase_ft_xC_fMRI_train[1:length(xC_fMRI_train)])

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
