% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseGen.R
\name{noiseGen}
\alias{noiseGen}
\title{noiseGen}
\usage{
noiseGen(data)
}
\arguments{
\item{data}{A data in the form of vector or of a matrix}
}
\value{
A updated version of data where noise is also included
}
\description{
This function is useful for generating noises for a given dataset
}
\details{
Noise-generating function to generate some synthetic predictors (X_new)
and test the ARIMA(2,0,1)(1,0,1)[12] model
}
\examples{
noiseGen( matrix(rep(2,9),3,3) )

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
