% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_intervals.R
\name{p_intervals}
\alias{p_intervals}
\title{Interval Probabilities for Chi-square Test}
\usage{
p_intervals(breaks)
}
\arguments{
\item{breaks}{Numeric vector of interval boundaries.}
}
\value{
A numeric vector of probabilities for each interval.
}
\description{
Computes theoretical probabilities over subintervals of [0,2]
for the custom distribution. These probabilities are used in
the Chi-squared goodness-of-fit test.
}
\details{
The probabilities are computed using the cumulative distribution function:
\deqn{
P(x_{i-1} \le X \le x_i) = F(x_i) - F(x_{i-1})
}

where:
\deqn{
F(x) = \frac{3}{2}x^2 - x^3 + \frac{3}{16}x^4
}
}
\examples{
breaks <- seq(0, 2, by = 0.2)
p_intervals(breaks)

}
