% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_wrapper.R
\name{pi_wrapper}
\alias{pi_wrapper}
\title{pi_wrapper}
\usage{
pi_wrapper(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A number that satifies one of the 4 cases specified below
}
\description{
Define a pi-wrapper calculator to ensure phase-arithmetic stays within [-pi : pi)
}
\details{
This function makes sure that phase-arithmetic is between [-pi, pi]
}
\examples{
pi_wrapper(pi/8 + 1.186667 )

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
