% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhaseDistributions.R
\name{plotPhaseDistributions}
\alias{plotPhaseDistributions}
\title{plotPhaseDistributions}
\usage{
plotPhaseDistributions(dataFT, dataColnames, size = 10, option = "ALL",
  select_feature = NULL, ...)
}
\arguments{
\item{dataFT}{data inherited from function "kSpaceTransform", K-space transformed dataset.}

\item{dataColnames}{feature names for the transformed dataset}

\item{size}{size of the text inside the plot}

\item{option}{two options avaliable: "All": to plot for all features, "Select": to select several features you wish to show}

\item{select_feature}{if "option" is setted as "Select", then put inside a sequence to indicate the features you wish to show with this function.}
}
\value{
This function returns a plot based on ggplot2 showing results of K-space transformation
}
\description{
Creating plot result for K-space transformation
}
\details{
This function provides a visualization method to show results of K-space transformation
This function can choose to generate all plots related to the K-space transformation.
Or we could choose features code to select features we like to see to generate K-space transformation result.
}
\examples{
require(ggplot2)
require(dplyr)
require(tidyr)

# Showing all results based on K-space transformation
## Make sure to change the height and width of your plot screen so that the plot can fit your monitor
plotPhaseDistributions(FT_Belgium, dataColnames)

# Only show a few features from the results. Here we choose to generate the first 6 results
plotPhaseDistributions(FT_Belgium, dataColnames,option = "Select",size = 10,select_feature = c(1,2,3,4,5,6))

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
