% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS_ARIMAX_remodel.R
\name{PS_ARIMAX_remodel}
\alias{PS_ARIMAX_remodel}
\title{PS_ARIMAX_remodel}
\usage{
PS_ARIMAX_remodel(X, Y, ts_Y_test, option = "Nil-Phase",
  result = "ARIMAX", rename_Y = "Y_GDP_Belgium")
}
\arguments{
\item{X}{inherited from function "homologousX_features", we can get parameter X by
"homologousX_features(country1,country2)$$X_Country1"}

\item{Y}{inherited from function "preprocess_ARIMA", we can get parameter Y by "preprocess_ARIMA$Y"}

\item{ts_Y_test}{time series vector, must be set from function "ts" and
must be created based on testing set}

\item{option}{choose the phase synthesis method used in the reconstruction or ARIMAX model.
Three options are valid: "Nil-Phase", "Swapped-Phase", "Random-Phase"}

\item{result}{choose the result shown by this function. Two options are valid:
"ARIMAX": this will show the ARIMAX model created by the chosen phase synthesis method.
"reconstruct": this will show the reconstruction procedure based on the chosen phase synthesis method,
but the ARIMAX model won't be built.}

\item{rename_Y}{rename the return result of Y}
}
\value{
If the "result" parameter is setted as "reconstruct", then the reconstruction result will be shown.
If setted as "ARIMAX", then the detailed ARIMAX model will be built.
}
\description{
ARIMAX Remodeling
}
\details{
Conduct three different phase synthesis methods and create ARIMAX model based on different method.
This function can reconstruct the dataset based on three phase synthesis methods:
Nil-Phase Synthesis, Swapped-Phase Synthesis and Random-Phase Synthesis.
Also it can construct different ARIMAX model based on the phase synthesis method chosen.
}
\examples{
require(DT)
# Show the reconstruction results based on "Nil-Phase" transformation
Nil_reconstruct<-PS_ARIMAX_remodel(X_Belgium,Y_Belgium,ts_Y_Belgium_test,result = "reconstruct")
## All different features under the "Nil-Phase"
datatable(data.frame(feature_name=Nil_reconstruct[[4]]),fillContainer = TRUE)
## All reconstruct results
datatable(data.frame(Nil_reconstruct[[6]]),fillContainer = TRUE)

# Example for showing the ARIMA model based on "Swapped-Phase" transformation
Swapped_ARIMAX<-PS_ARIMAX_remodel(X_Belgium,Y_Belgium,ts_Y_Belgium_test,option = "Swapped-Phase",result = "ARIMAX")
## ARIMA information (AR & MA)
Swapped_ARIMAX$AR_MA
## Model & time series information
Swapped_ARIMAX$ARIMAX_model
## Information for most important features
Swapped_ARIMAX$feature_effect
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
