% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomfun2.R
\name{randomfun2}
\alias{randomfun2}
\title{Random pseudo-numbers generator via Newton-Raphson}
\usage{
randomfun2(n, tol = 1e-08, max_iter = 100)
}
\arguments{
\item{n}{Integer. Length of the pseudo-random vector to be generated.}

\item{tol}{Numeric. Tolerance for convergence (default 1e-8).}

\item{max_iter}{Integer. Maximum number of iterations (default 100).}
}
\value{
A numeric vector of length \code{n} with random values.
}
\description{
Generates pseudo-random numbers from a continuous distribution defined by
a quartic cumulative distribution function (CDF), using the inverse transform
method resovled by Newton-Raphson algorithm.
}
\details{
This function applies the inverse transform method:
\deqn{F(x) = \tfrac{3}{2}x^2 - x^3 + \tfrac{3}{16}x^4, \quad x \in [0,2]}

The equation \eqn{F(x) - u = 0} is solved using the Newton-Raphson
iterative scheme:
\deqn{x_{n+1} = x_n - \frac{F(x_n) - u}{f(x_n)}}
}
\examples{
# Generate 100 random values
randomfun2(100)

}
