% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Calculate the sample mean}
\usage{
sample_mean(x)
}
\arguments{
\item{x}{a vector of numbers of length N}
}
\value{
the sample mean of x
}
\description{
Calculate the sample mean
}
\details{
This function calculate the sample mean
}
\examples{
a <- c(1,2,3,4,5)
sample_mean(a)

}
