% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spacetime_Kime_Rank.R
\name{Spacetime_Kime_Rank}
\alias{Spacetime_Kime_Rank}
\title{Spacetime_Kime_Rank}
\usage{
Spacetime_Kime_Rank(X, Y, signal = "Weak", phase = "Normal",
  result = 1, Seed = 4321, lam = "min")
}
\arguments{
\item{X}{a data frame X containing all predictors. X needs to contain features both from ARIMA time series model and
also features for the external regressors(i.e. the "xreg" part).}

\item{Y}{a vector Y containing information of the response result.}

\item{signal}{two options for this parameter: "Weak" and "Strong". Set to "weak" will analyse the model only based on 
ARIMA features (i.e. features with AR and MA parts). Set to "strong" will analyse model based on all features}

\item{phase}{three options: "Normal", "Nil" and "Swap". Using "Normal" will not conduct SpaceKime transformation.
"Nil" will conduct "Nil-Phase" transformation and "Swap" will conduct "Swapped-Phase" transformation on the data.}

\item{result}{There are generally 5 choices to show 5 different results. If result=1, then the MSE plot of LASSO
cross-validation will be shown. If result=2 then the coefficient of the top 9 sailent features under LASSO penalty
will be shown. If result=3 then the ranking comparison with the observed value against the predicted value (under
different SpaceKime transformation results) will be shown. Also the correlation will also be shown. If result=4 a plot
similiar to the result of 3 will be shown. But this time the plot will also include the information about whether a 
country is top 30 country or not. And the name of different countries will also be shown on this plot to give readers
a much clearer view. If result=5, then a list will be shown, this list contains dimension of predictors used in our model,
information about first 9 highest features, correlation between predicted and true results, all features importance
comparison, and other results based on the choice of another parameter "lam".
Moreover, there are another two options, result=6 and result=7. These two options will only work when we choose phase
parameter as "Nil". Result=6 will show a plot based on phase distribution.}
}
\description{
Generate numerous LASSO regression analysis results based on 1. "Weak" or "Strong" signals 2. different
Space Kime transformation methods
}
\details{
This function \code{Spacetime_Kime_Rank} is a function that takes in a data frame X and a vector Y. X is the 
data fram for all predicor variables and Y is the vector for the response variable. Then the function will base on two
other parameters "signal" and "phase" to determine which features will be used on creating ARIMAX models and also 
whether SpaceKime transformation methods will be applied on X and Y. It will generate numerous results. And the result
that is returned eventually is deterimined on parameter "result". More details can be viewed on the parameters part.
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
